+++ dhis-2/dhis-services/dhis-service-setting/src/main/java/org/hisp/dhis/setting/SystemSetting.java
@@ -39,7 +39,7 @@ import org.hisp.dhis.common.BaseIdentifiableObject;
 public class SystemSetting extends BaseIdentifiableObject
     implements Serializable
 {
-    private int id;
+    private long id;
 
     private String name;
 
@@ -66,12 +66,12 @@ public class SystemSetting extends BaseIdentifiableObject
     // Getters and setters
     // -------------------------------------------------------------------------
 
-    public int getId()
+    public long getId()
     {
         return id;
     }
 
-    public void setId( int id )
+    public void setId( long id )
     {
         this.id = id;
     }
